+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 -   Override dbt-core `default__type_string()` macro to use Dremio Supported VARCHAR instead of the default string. ([#80](https://github.com/dremio/dbt-dremio/pull/80))
 
--   Change \_populate_job_results() to have an optional page_limit argument with default set to 100 (Dremio's default).
+-   Change \_populate_job_results() to have an optional row_limit argument with default set to 100 (Dremio's default).
 
 -   Implement pagination in \_populate_job_results()